requirements.txt

# Thư viện phục vụ xử lý dữ liệu của bạn
apache-airflow==3.1.5
pandas>=2.1.2,<2.2
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
kaggle==1.6.14

